+++ native/common/include/jp_chartype.h
@@ -82,6 +82,10 @@ public:
 	virtual void copyElements(JPJavaFrame &frame,
 			jarray a, jsize start, jsize len,
 			void* memory, int offset) override;
+
+	virtual PyObject *newMultiArray(JPJavaFrame &frame,
+			JPPyBuffer &buffer, int subs, int base, jobject dims) override;
+
 } ;
 
 #endif // _JP-CHAR_TYPE_H_