+++ native/common/jp_booleantype.cpp
@@ -278,3 +278,15 @@ void JPBooleanType::copyElements(JPJavaFrame &frame, jarray a, jsize start, jsiz
 	jboolean* b = (jboolean*) ((char*) memory + offset);
 	frame.GetBooleanArrayRegion((jbooleanArray) a, start, len, b);
 }
+
+static void pack(jboolean* d, jvalue v)
+{
+	*d = v.z;
+}
+
+PyObject *JPBooleanType::newMultiArray(JPJavaFrame &frame, JPPyBuffer &buffer, int subs, int base, jobject dims)
+{
+	return convertMultiArray<type_t>(
+			frame, this, &pack, "z",
+			buffer, subs, base, dims);
+}