+++ native/common/jp_inttype.cpp
@@ -308,3 +308,15 @@ void JPIntType::copyElements(JPJavaFrame &frame, jarray a, jsize start, jsize le
 	jint* b = (jint*) ((char*) memory + offset);
 	frame.GetIntArrayRegion((jintArray) a, start, len, b);
 }
+
+static void pack(jint* d, jvalue v)
+{
+	*d = v.i;
+}
+
+PyObject *JPIntType::newMultiArray(JPJavaFrame &frame, JPPyBuffer &buffer, int subs, int base, jobject dims)
+{
+	return convertMultiArray<type_t>(
+			frame, this, &pack, "i",
+			buffer, subs, base, dims);
+}