+++ doc/CHANGELOG.rst
@@ -4,10 +4,14 @@ Changelog
 This changelog *only* contains changes from the *first* pypi release (0.5.4.3) onwards.
 
 - **Next version - unreleased**
+
   - java.util.Map conforms to Python abc.Mapping API.
 
   - float properly follows Java rules for conversion from double.
     floats outside of range map to inf and -inf.
+
+  - Add support for direct conversion of multi-dimensional primitive arrays
+    with ``JArray.of(array, [dtype=type])``
 
 - **0.7.2 - 2-28-2019**